+++ phase1/src/group/trade/TradeManager.java
@@ -44,8 +44,8 @@ public class TradeManager {
      * @param prevMeeting The trade ID of the previous meeting
      * @return A response object of the representation of the trade or a description of why creation failed
      */
-    public Response createTrade(long user1, long user2, long item1, long item2, Boolean isPermanent,
-                                LocalDateTime dateAndTime, String location, Long prevMeeting) {
+    public Response createTrade(int user1, int user2, int item1, int item2, Boolean isPermanent,
+                                LocalDateTime dateAndTime, String location, Integer prevMeeting) {
         // Get users from Repository
         PersonalUser trader1 = userRepository.get(user1);
         PersonalUser trader2 = userRepository.get(user2);
@@ -68,7 +68,7 @@ public class TradeManager {
      * @param location The location of where this trade will take place
      * @return A response object of the representation of the trade or a description of why creation failed
      */
-    public Response createTrade(long user1, long user2, long item1, long item2, Boolean isPermanent,
+    public Response createTrade(int user1, int user2, int item1, int item2, Boolean isPermanent,
                                 LocalDateTime dateAndTime, String location){
         return createTrade(user1, user2, item1, item2, isPermanent, dateAndTime, location, null);
     }
